% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{depl_check}
\alias{depl_check}
\title{Check R source dependencies.}
\usage{
depl_check(source_path = ".")
}
\arguments{
\item{source_path}{a file or directory path.}
}
\value{
A list of dataframes containing package metadata. Convenient for diagnostics or further processing.
}
\description{
`depl_check` analyses R source for dependency references:
`library(<package>)`, `require(<package>)`, `<package>::func`,
`pload(<package>, <package>)`. The dependencies are then located in either the
local library, CRAN, or GitHub. Installed versions are compared with repository versions,
with CRAN taking precendence over GitHub (Even if a local package was installed from GitHub).
}
\details{
A report is output to the console containing information about missing and out of date dependencies.
If there are any missing or out of date, a prompt is made to automatically update and install dependencies.
}
\examples{
\dontrun{
depl_check("~/repos/my_analysis")
depl_check("./report.Rmd")
}

}
