% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{depl_check_addin}
\alias{depl_check_addin}
\title{Check R dependencies in the currently active Rstudio pane.}
\usage{
depl_check_addin()
}
\value{
A list of dataframes containing package metadata. Convenient for diagnostics or further processing.
}
\description{
See `depl_check()` for details.
}
