% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_libs.R
\name{find_doc_libs}
\alias{find_doc_libs}
\title{Find libraries referenced in a document containing R code}
\usage{
find_doc_libs(doc)
}
\arguments{
\item{doc}{a character vector or list of character vectors containing R code.}
}
\value{
A vector of characters containing library references present in the Code
}
\description{
Find libraries referenced in a document containing R code
}
\examples{
\dontrun{
find_doc_libs(c("library(datapasta)\\n", "devtools::check()\\n", "#require(shist)"))
}
}
